slight restructuring of project
