+++ r8te.py
@@ -1,0 +1,7 @@
+import botHandler
+
+
+if __name__ == '__main__':
+    botHandler.run_discord_bot()
+
+
